+++ database/migrations/2023_09_27_080912_create_nodal_officers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('nodal_officer_id');
             $table->string('officer_name',50);
             $table->string('officer_designation',100);
-            $table->unsignedInteger('officer_contact')->unique();
+            $table->unsignedBigInteger('officer_contact')->unique();
             $table->unsignedBigInteger('relief_camp_id');
             $table->boolean('active_status')->default(true);
             $table->timestamps();
